+++ src/app/loadout-drawer/auto-loadouts.ts
@@ -1,5 +1,6 @@
 import { t } from 'app/i18next-t';
 import { D1BucketHashes } from 'app/search/d1-known-values';
+import { D2ItemTiers } from 'app/search/d2-known-values';
 import { ItemFilter } from 'app/search/filter-types';
 import { isD1Item, itemCanBeEquippedBy } from 'app/utils/item-utils';
 import { BucketHashes } from 'data/d2/generated-enums';
@@ -46,7 +47,7 @@ export function itemLevelingLoadout(allItems: DimItem[], store: DimStore): Loado
       value += 500;
     }
 
-    value += ['Common', 'Uncommon', 'Rare', 'Legendary', 'Exotic'].indexOf(item.tier) * 10;
+    value += D2ItemTiers[item.tier] * 10;
 
     // Choose the item w/ the highest XP
     if (isD1Item(item) && item.talentGrid) {