+++ src/app/search/specialty-modslots.ts
@@ -13,6 +13,8 @@ export interface ModSocketMetadata {
   emptyModSocketHash: number;
   /** so you can look these entries up if all you're given is an "Empty Mod Slot" plug item */
   emptyModSocketHashes: number[];
+  /** the year is 2022. the raid is Vow of the Disciple. bungie forgot to give raid mods a itemTypeDisplayName. let's use this Activity name instead. */
+  modGroupNameOverrideActivityHash?: number;
 }
 
 const legacyCompatibleTags = [
@@ -55,6 +57,7 @@ export const modTypeTagByPlugCategoryHash = {
   [PlugCategoryHashes.EnhancementsSeasonV490]: 'chargedwithlight',
   [PlugCategoryHashes.EnhancementsRaidDescent]: 'deepstonecrypt',
   [PlugCategoryHashes.EnhancementsRaidV520]: 'vaultofglass',
+  [PlugCategoryHashes.EnhancementsRaidV600]: 'vowofthedisciple',
   [PlugCategoryHashes.EnhancementsSeasonV500]: 'combat',
 };
 
@@ -123,6 +126,15 @@ const modSocketMetadata: ModSocketMetadata[] = [
     compatiblePlugCategoryHashes: [PlugCategoryHashes.EnhancementsRaidV520],
     emptyModSocketHashes: [3738398030],
     emptyModSocketHash: 3738398030,
+  },
+  {
+    slotTag: 'vowofthedisciple',
+    compatibleModTags: ['vowofthedisciple'],
+    socketTypeHashes: [2381877427],
+    compatiblePlugCategoryHashes: [PlugCategoryHashes.EnhancementsRaidV600],
+    emptyModSocketHashes: [2447143568],
+    emptyModSocketHash: 2447143568,
+    modGroupNameOverrideActivityHash: 2906950631, // Activity "Vow of the Disciple"
   },
   {
     slotTag: 'combatstyle',